Add CSS import rewriter tests
